Moved script classes from base to script package
