+++ expansion/src/main/java/com/extendedclip/papi/expansion/javascript/commands/DebugCommand.java
@@ -2,7 +2,7 @@ package com.extendedclip.papi.expansion.javascript.commands;
 
 import com.extendedclip.papi.expansion.javascript.ExpansionUtils;
 import com.extendedclip.papi.expansion.javascript.JavascriptPlaceholder;
-import com.extendedclip.papi.expansion.javascript.ScriptRegistry;
+import com.extendedclip.papi.expansion.javascript.script.ScriptRegistry;
 import com.extendedclip.papi.expansion.javascript.commands.router.ExpansionCommand;
 import org.bukkit.command.CommandSender;
 import org.bukkit.util.StringUtil;
