+++ expansion/src/main/java/com/extendedclip/papi/expansion/javascript/script/ConfigurationScriptLoader.java
@@ -1,5 +1,6 @@
-package com.extendedclip.papi.expansion.javascript;
+package com.extendedclip.papi.expansion.javascript.script;
 
+import com.extendedclip.papi.expansion.javascript.JavascriptPlaceholder;
 import com.extendedclip.papi.expansion.javascript.config.ScriptConfiguration;
 import com.extendedclip.papi.expansion.javascript.evaluator.ScriptEvaluatorFactory;
